+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - [bug] wheel scroll bug on web
 - [bug] pointer scroll support on web
+- [bug] drag and drop on calendar not full width
+- [improve] change vertical scroll during drag and drop
 - [documentation] web demo
 
 # [2.0.1 - 16 Jan 2025](https://github.com/pickywawa/infinite_calendar_view/tree/2.0.1)